requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-crontab
newsapi-python
azure-communication-identity
django-cors-headers
mysqlclient==2.0.0
stripe==5.0.0
aiohttp==3.8.3
requests==2.28.1
pytz==2022.7
django-environ
pillow
GDAL==3.2.2.1
gunicorn
